+++ packages/expo-cli/commands/install.js
@@ -5,15 +5,17 @@ const { join } = require('path')
 const install = require('../lib/install')
 const { onCancel } = require('../lib/utils')
 const { blue } = require('kleur')
+const semver = require('semver')
 
 module.exports = async (argv, globalOpts) => {
   const projectRoot = globalOpts['project-root']
   const alreadyInstalled = await checkManifest(projectRoot)
   const isWanted = await confirmWanted(alreadyInstalled, projectRoot)
   if (isWanted) {
+    const version = await selectVersion(projectRoot)
     const tool = await withTool(projectRoot)
     console.log(blue(`> Installing @bugsnag/expo with ${tool}. This could take a while!`))
-    await install(tool)
+    await install(tool, version, projectRoot)
   }
 }
 
@@ -60,6 +62,31 @@ const checkManifest = async (dir) => {
     return allDeps.includes('@bugsnag/expo')
   } catch (e) {
     throw new Error('Could not load package.json. Is this the project root?')
+  }
+}
+
+const selectVersion = async (dir) => {
+  try {
+    const pkg = JSON.parse(await promisify(readFile)(join(dir, 'package.json'), 'utf8'))
+    const expoVersion = pkg.dependencies.expo
+    const isPre33 = (expoVersion && !semver.gte(semver.minVersion(expoVersion), '33.0.0'))
+    const { version } = await prompts({
+      type: 'text',
+      name: 'version',
+      message: isPre33
+        ? 'It looks like you’re using a version of Expo SDK <33. The latest version of Bugsnag works with SDK >= 33 so it’s recommended that you install the last version of Bugsnag that supported your Expo version: v6.3.0'
+        : 'If you want the latest version of @bugsnag/expo hit enter, otherwise type the version you want',
+      initial: isPre33 ? '6.3.0' : 'latest',
+      validate: str => {
+        if (str === 'latest') return true
+        if (semver.valid(str)) return true
+        if (semver.validRange(str)) return true
+        return 'Version must be: a valid semver version/range or "latest"'
+      }
+    }, { onCancel })
+    return version
+  } catch (e) {
+    throw new Error(`Could not detect Expo version in package.json: ${e.message}`)
   }
 }