+++ packages/plugin-inline-script-content/inline-script-content.js
@@ -102,22 +102,25 @@ module.exports = {
       'Notification', 'SVGElementInstance', 'Screen', 'TextTrack', 'TextTrackCue', 'TextTrackList',
       'WebSocket', 'WebSocketWorker', 'Worker', 'XMLHttpRequest', 'XMLHttpRequestEventTarget', 'XMLHttpRequestUpload'
     ], o => {
-      if (!win[o] || !win[o].prototype || typeof win[o].prototype.addEventListener !== 'function') return
+      if (!win[o] || !win[o].prototype || !win[o].prototype.hasOwnProperty || !win[o].prototype.hasOwnProperty('addEventListener')) return
       __proxy(win[o].prototype, 'addEventListener', original =>
         __traceOriginalScript(original, eventTargetCallbackAccessor)
       )
       __proxy(win[o].prototype, 'removeEventListener', original =>
-        __traceOriginalScript(original, eventTargetCallbackAccessor)
+        __traceOriginalScript(original, eventTargetCallbackAccessor, true)
       )
     })
 
-    function __traceOriginalScript (fn, callbackAccessor) {
+    function __traceOriginalScript (fn, callbackAccessor, alsoCallOriginal = false) {
       return function () {
-        var args = Array.prototype.slice.call(arguments)
-        var cba = callbackAccessor(args)
-        var cb = cba.get()
-        if (typeof cb !== 'function') return fn.apply(this, args)
+        // this is required for removeEventListener to remove anything added with
+        // addEventListener before the functions started being wrapped by Bugsnag
+        const args = Array.prototype.slice.call(arguments)
         try {
+          const cba = callbackAccessor(args)
+          const cb = cba.get()
+          if (alsoCallOriginal) fn.apply(this, args)
+          if (typeof cb !== 'function') return fn.apply(this, args)
           if (cb.__trace__) {
             cba.replace(cb.__trace__)
           } else {
@@ -142,6 +145,7 @@ module.exports = {
         } catch (e) {
           // swallow these errors on Selenium:
           // Permission denied to access property '__trace__'
+          // WebDriverException: Message: Permission denied to access property "handleEvent"
         }
         // IE8 doesn't let you call .apply() on setTimeout/setInterval
         if (fn.apply) return fn.apply(this, args)