+++ test/scripts/create_package.json_tar.sh
@@ -1,0 +1,5 @@
+#!
+
+rm -f min_packages.tar
+
+find packages/** -maxdepth 2 -type f \( -name package.json -or -name package-lock.json \) | tar --mtime='1970-01-01' -cf - -T - | gzip --no-name > min_packages.tar